Add AAAAXY, a platformer game taking place in impossible spaces.
This game is notable for doing a thing few other 2D games do - it
takes place in a world that globally defies the rules of geometry
as we know it, messing with your sense of orientation. This is
achieved by means of seamless portals placed in the game world (like
many have already done in 3D first person games, where this feature
would be trivial).

It runs as native application on Linux, Windows, macOS, "mostly native"
on FreeBSD (it currently requires the ALSA library as a wrapper to OSS,
which is in the ports collection), and is also available for mobile
operating systems and HTML5.

It was made using Go and Ebitengine (basically a platform independent
graphics/sound/main loop library for Go that does not restrict game
engine design and thus allowed me to do these things).